+++ docs/history.rst
@@ -5,6 +5,7 @@ Version History
 --------------------------------
 - Add a transcoder for `application/x-www-formurlencoded`_
 - Add support for encoding :class:`decimal.Decimal`
+- Add support for encoding :func:`dataclasses.dataclass` decorated classes
 - Add type annotations (see :ref:`type-info`)
 - Return a "406 Not Acceptable" if the :http:header:`Accept` header values cannot be matched
   and there is no default content type configured